Move PlacesNearby viewmodel logic into separate file for sharing
We need to implement PlacesNearby in AddWaypoints, so share the code
from the viewmodel.
