+++ app/src/main/java/org/scottishtecharmy/soundscape/screens/home/placesnearby/PlacesNearbyScreen.kt
@@ -24,7 +24,7 @@ fun PlacesNearbyScreenVM(
     modifier: Modifier = Modifier,
     viewModel: PlacesNearbyViewModel =  hiltViewModel(),
 ) {
-    val uiState by viewModel.uiState.collectAsStateWithLifecycle()
+    val uiState by viewModel.logic.uiState.collectAsStateWithLifecycle()
     PlacesNearbyScreen(
         homeNavController,
         uiState,
